Evolution
I refactored the metadata into a separate component; I'm hoping we can re-use
this as a flyover on the comparison page, but haven't quite figured out how to
achieve that.

I'm "stacking" the metrics summary and graph, and I moved the graph legends
underneath the graph, horizontally, to save space.