Report primary period better, and use it in the UI
The `samples` API now returns the associated iteration primary-period along
with the primary-metric. The `periods` API now identifies the iteration
primary-metric only for the primary period. The UI actions use this new
info to avoid trying to summarize or graph the primary-metric in non-primary
periods where it won't have a value.